+++ src/main/java/ntnu/idatt/boco/service/UserService.java
@@ -15,6 +15,6 @@ public interface UserService {
     Role saveRole(Role role);
     void addRoleToUser(String username, String roleName);
     User getUser(String username);
-    User getUserById(Long id);
+    User getUserById(Integer id);
     List<User> getUsers();
 }
